[feat] viewModel로 giftId 전달

+++ app/src/main/java/org/gdsc/donut/data/remote/response/home/ResponseHomeReceiverBoxItem.kt
@@ -14,7 +14,7 @@ data class ResponseHomeReceiverBoxItemData(
 )
 
 data class ResponseHomeReceiverGift(
-    val giftId: Int,
+    val giftId: Long,
     val product: String,
     val price: Int,
     val dueDate: String,
